Fruit ninja game, also known as fruit-slicing game which is easy to play. Fruit ninja game is popular among children.  The objective of this project is to build a fruit ninja game with python. This game is built with the help of pygame module and basic concept of python.  In this game, the user has to cut the fruits by touching the mouse on fruits. There are also bombs with fruits. If the mouse touches more than three bombs then the game will be over.
Fruit ninja game, also known as fruit-slicing game which is easy to play. Fruit ninja game is popular among children.

The objective of this project is to build a fruit ninja game with python. This game is built with the help of pygame module and basic concept of python.

In this game, the user has to cut the fruits by touching the mouse on fruits. There are also bombs with fruits. If the mouse touches more than three bombs then the game will be over.

Project Summary
We have successfully created the fruit ninja game python project. We used the popular pygame library. We learned how to randomly generate fruits in particular positions. I hope you enjoyed building this game project.